AOI抽成模块
