+++ Apps/Hotfix/Module/AOI/AOIManagerComponentSystem.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace ET.Server
+namespace ET
 {
     [FriendClass(typeof(AOIManagerComponent))]
     [FriendClass(typeof(AOIEntity))]
