+++ Apps/Model/Module/AOI/AOIEventType.cs
@@ -1,4 +1,4 @@
-﻿namespace ET.Server
+﻿namespace ET
 {
 	namespace EventType
 	{
