+++ Apps/Model/Module/AOI/Cell.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace ET.Server
+namespace ET
 {
     public class Cell: Entity, IAwake, IDestroy
     {
