+++ Unity/Codes/Model/Module/Move/MoveEventType.cs
@@ -1,0 +1,13 @@
+﻿namespace ET
+{
+    namespace EventType
+    {
+        public struct MoveStart
+        {
+        }
+
+        public struct MoveStop
+        {
+        }
+    }
+}
