+++ Unity/Codes/Model/Module/Unit/UnitEventType.cs
@@ -1,0 +1,16 @@
+﻿using UnityEngine;
+
+namespace ET
+{
+    namespace EventType
+    {
+        public struct ChangePosition
+        {
+            public Vector3 OldPos;
+        }
+
+        public struct ChangeRotation
+        {
+        }
+    }
+}
