+++ Unity/Codes/Model/Share/ETCancelationTokenHelper.cs
